% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cvdata}
\alias{cvdata}
\title{Example CV data}
\format{
A list holding lists of CV data.
}
\usage{
data(cvdata)
}
\description{
Example CV data in the format expected by the \code{qmdcv} package.
}
\keyword{datasets}
