% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example CV data}
\format{
A list.
}
\usage{
data(example_data, package = "qmdcv")
}
\description{
Example CV data in the format expected by the \code{qmdcv} package.
}
\keyword{datasets}
