% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.R
\name{insert_pubs}
\alias{insert_pubs}
\title{Insert publications into your CV}
\usage{
insert_pubs(data)
}
\arguments{
\item{data}{A list in the format expected by the \code{qmdcv} package. See
\code{data(example_data)} for an example.}
}
\value{
Markdown text to be rendered with Quarto.
}
\description{
\code{insert_pubs()} converts a structured list into Quarto-style
Markdown text to be rendered as a list of publications.
}
\details{
This function currently only works for journal articles.

Note: the code chunk in which this function is run should be given
the option \code{output: asis}
}
\examples{
data(example_data)
pubs <- example_data$publications$articles
insert_pubs(pubs)
}
\seealso{
\code{\link[=insert]{insert()}}, \code{\link[=insert_talks]{insert_talks()}}, \code{\link[=insert_list]{insert_list()}}
}
