% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.R
\name{insert_pubs_web}
\alias{insert_pubs_web}
\title{Insert publications into your web page}
\usage{
insert_pubs_web(data)
}
\arguments{
\item{data}{A list of one or more lists, where each sub-list contains
publication data. See \code{data(example_data)} for an example.}
}
\value{
Markdown text to be rendered with Quarto.
}
\description{
\code{insert_pubs_web()} converts a list of publication data into
Quarto-style Markdown text, as does \code{\link[=insert_pubs]{insert_pubs()}}, but with additional
formatting for web pages. This function only supports journal articles at
present.
}
\details{
The following sub-list elements are recognized:
\itemize{
\item \code{authors} (string or vector)
\item \code{year} (string)
\item \code{title} (string)
\item \code{publication} (string)
\item \code{volume} (string)
\item \code{pages} (string)
\item \code{doi} (string)
\item \code{data} (string)
\item \code{code} (string)
}
}
\note{
This function uses the
\href{https://github.com/quarto-ext/fontawesome}{Font Awesome Extension} for
Quarto. Make sure to install this extension to the project root directory.
}
\examples{
\dontrun{
data(example_data)
pubs <- example_data$publications$articles
insert_pubs_web(pubs)
}
}
\seealso{
\code{\link[=insert_pubs]{insert_pubs()}}
}
