% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.R
\name{insert_talks}
\alias{insert_talks}
\title{Insert talks/presentations into your CV}
\usage{
insert_talks(data)
}
\arguments{
\item{data}{A list of one or more lists, where each sub-list contains
talk/presentation data. See \code{data(cvdata)} for an example.}
}
\value{
Markdown text to be rendered with Quarto.
}
\description{
\code{insert_talks()} converts a list of talk/presentation data into
Quarto-style Markdown text.
}
\details{
The following sub-list elements are recognized:
\itemize{
\item \code{authors} (string or vector)
\item \code{date} (string)
\item \code{title} (string)
\item \code{context} (string)
}
}
\examples{
data(cvdata)
talks <- cvdata$talks
insert_talks(talks)
}
