% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcv_insert.R
\name{qcv_insert_list}
\alias{qcv_insert_list}
\title{Insert a list into your CV}
\usage{
qcv_insert_list(data, type = "u")
}
\arguments{
\item{data}{A list in the format expected by the \code{qcvutils} package. See
\code{data(qcvexample)} for an example.}

\item{type}{Type of list to insert. Choose from among:
\itemize{
\item \code{"u"}: unordered list
\item \code{"1"}: numbered list
\item \code{"a"}: alphabetical list
\item \code{"n"}: no bullets
}}
}
\value{
Markdown text to be rendered with Quarto.
}
\description{
\code{qcv_insert_list} converts a structured list (typically imported
from a YAML file) into Quarto-style markdown text (including HTML/CSS
where necessary) to be rendered as a CV section. Note that the code chunk in
which this function is run should be given the \code{output: asis} option.
}
\examples{
data(qcvexample)
teaching <- qcvexample$teaching
qcv_insert_list(teaching)

}
\seealso{
\url{qcv_insert}, \url{qcv_insert_pubs}, \url{qcv_insert_talks}
}
