% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcv_insert.R
\name{qcv_insert_pubs}
\alias{qcv_insert_pubs}
\title{Insert publications into your CV}
\usage{
qcv_insert_pubs(data)
}
\arguments{
\item{data}{A list in the format expected by the \code{qcvutils} package. See
\code{data(qcvexample)} for an example.}
}
\value{
Markdown text to be rendered with Quarto.
}
\description{
\code{qcv_insert_pubs} converts a structured list (typically
imported from a YAML file) into Quarto-style markdown text (including HTML/CSS
where necessary) to be rendered as a list of publications. Note that the code
chunk in which this function is run should be given the \code{output: asis} option.
}
\details{
This function currently only works for journal articles.
}
\examples{
data(qcvexample)
pubs <- qcvexample$publications$articles
qcv_insert_pubs(pubs)

}
\seealso{
\url{qcv_insert}
}
