% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcv_insert.R
\name{qcv_insert_talks}
\alias{qcv_insert_talks}
\title{Insert talks into your CV}
\usage{
qcv_insert_talks(data)
}
\arguments{
\item{data}{A list in the format expected by the \code{qcvutils} package. See
\code{data(qcvexample)} for an example.}
}
\value{
Markdown text to be rendered with Quarto.
}
\description{
\code{qcv_insert_talks()} converts a structured list into Quarto-style
markdown text to be rendered as a list of talks.
}
\details{
Note: the code chunk in which this function is run should be given
the option \code{output: asis}
}
\examples{
data(qcvexample)
talks <- qcvexample$talks
qcv_insert_talks(talks)
}
\seealso{
\code{\link[=qcv_insert]{qcv_insert()}}, \code{\link[=qcv_insert_pubs]{qcv_insert_pubs()}}, \code{\link[=qcv_insert_list]{qcv_insert_list()}}
}
