% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qcvexample}
\alias{qcvexample}
\title{Example CV data}
\format{
A list.
}
\usage{
data(qcvexample)
}
\description{
A list containing example CV data in the format expected by the \code{qcvutils}
package.
}
\keyword{datasets}
