% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{read_example}
\alias{read_example}
\title{Read example CV data}
\usage{
read_example(...)
}
\arguments{
\item{...}{Potential arguments passed to \code{\link[yaml:read_yaml]{yaml::read_yaml()}}.}
}
\value{
A list containing example CV data.
}
\description{
Read example CV data
}
