% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{write_example}
\alias{write_example}
\title{Write example CV data}
\usage{
write_example(to = "", ...)
}
\arguments{
\item{to}{Path to a single destination file or directory. "" indicates output
to the console.}

\item{...}{Potential arguments passed to \code{\link[=file.copy]{file.copy()}}.}
}
\value{
When outputting to the console: none (invisible \code{NULL}).

When writing to a file or directory: \code{TRUE} or \code{FALSE} indicating whether
the operation succeeded.
}
\description{
Write example CV data
}
