% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.R
\name{insert}
\alias{insert}
\title{Insert entries into your CV}
\usage{
insert(x)
}
\arguments{
\item{x}{A YAML-style list of CV entries. The following elements are
recognized for each entry:
\itemize{
\item \code{title} (string)
\item \code{start} (string)
\item \code{end} (string)
\item \code{years} (string or vector)
\item \code{details} (string or vector)
\item \code{notes} (string or vector)
}}
}
\value{
None (invisible \code{NULL}).
}
\description{
Convert a list of CV entries into Quarto-style Markdown text.
}
\examples{
data(cvdata)
insert(cvdata$education)
}
