% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.R
\name{insert}
\alias{insert}
\title{Insert entries into your CV}
\usage{
insert(data)
}
\arguments{
\item{data}{A list in the format expected by the \code{qmdcv} package. See
\code{data(example_data)} for an example.}
}
\value{
Markdown text to be rendered with Quarto.
}
\description{
\code{insert()} converts a structured list into Quarto-style
Markdown text to be rendered as a CV section.
}
\details{
Note: the code chunk in which this function is run should be given
the option \code{output: asis}
}
\examples{
data(example_data)
edu <- example_data$education
insert(edu)
}
\seealso{
\code{\link[=insert_pubs]{insert_pubs()}}, \code{\link[=insert_talks]{insert_talks()}}, \code{\link[=insert_list]{insert_list()}}
}
