% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.R
\name{insert_list}
\alias{insert_list}
\title{Insert a list into a CV}
\usage{
insert_list(x, type = "u")
}
\arguments{
\item{x}{A YAML-style list of CV entries. The following elements are
recognized for each entry:
\itemize{
\item \code{title} (string)
\item \code{start} (string)
\item \code{end} (string)
\item \code{years} (string or vector)
\item \code{notes} (string or vector)
}}

\item{type}{The type of list to insert. Choose from among:
\itemize{
\item \code{"u"}: unordered
\item \code{"1"}: numbered
\item \code{"a"}: alphabetical
\item \code{"n"}: no bullets
}}
}
\value{
None (invisible \code{NULL}).
}
\description{
Convert a list of CV entries into a Quarto-style Markdown list.
}
\note{
This function can produce suboptimal results upon render when the
supplied \code{title} and/or \code{details} are long AND dates (\code{start}, \code{end},
\code{years}) are supplied. When using this function, it is safest to supply only
short \code{title} and \code{details}. If longer \code{title} or \code{details} are desired,
consider using \code{\link[=insert]{insert()}}.
}
\examples{
cvdata <- read_example()
insert_list(cvdata$teaching)
}
