% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.R
\name{insert_list}
\alias{insert_list}
\title{Insert a list into your CV}
\usage{
insert_list(data, type = "u")
}
\arguments{
\item{data}{A list in the format expected by the \code{qmdcv} package. See
\code{data(example_data)} for an example.}

\item{type}{Type of list to insert. Choose from among:
\itemize{
\item \code{"u"}: unordered list
\item \code{"1"}: numbered list
\item \code{"a"}: alphabetical list
\item \code{"n"}: no bullets
}}
}
\value{
Markdown text to be rendered with Quarto.
}
\description{
\code{insert_list()} converts a structured list into Quarto-style
Markdown text to be rendered as a list.
}
\details{
This function is currently known to provide undesirable output when
supplied \code{title} and/or \code{details} are long AND \code{start}/\code{end}/\code{years} are
provided. When using this function, one should ideally supply only short
\code{title} and \code{details}. If longer \code{title} and/or \code{details} are needed,
consider using \code{\link[=insert]{insert()}}.

Note: the code chunk in which this function is run should be given
the option \code{output: asis}
}
\examples{
data(example_data)
teaching <- example_data$teaching
insert_list(teaching)
}
\seealso{
\code{\link[=insert]{insert()}}, \code{\link[=insert_pubs]{insert_pubs()}}, \code{\link[=insert_talks]{insert_talks()}}
}
