% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.R
\name{insert_publications}
\alias{insert_publications}
\title{Insert publications into your CV}
\usage{
insert_publications(x)
}
\arguments{
\item{x}{A YAML-style list of publications. The following elements are
recognized for each item:
\itemize{
\item \code{citation} (string)
\item \code{notes} (string or vector)
}}
}
\value{
None (invisible \code{NULL}).
}
\description{
Convert a list of publications into a Quarto-style Markdown list.
}
\examples{
data(cvdata)
insert_publications(cvdata$publications)
}
