% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.R
\name{insert_pubs}
\alias{insert_pubs}
\title{Insert publications into your CV}
\usage{
insert_pubs(data)
}
\arguments{
\item{data}{A list of one or more lists, where each sub-list contains
publication data. See \code{data(cvdata)} for an example.}
}
\value{
Markdown text to be rendered with Quarto.
}
\description{
\code{insert_pubs()} converts a list of publication data into
Quarto-style Markdown text. This function only supports journal articles at
present.
}
\details{
The following sub-list elements are recognized:
\itemize{
\item \code{authors} (string or vector)
\item \code{year} (string)
\item \code{title} (string)
\item \code{publication} (string)
\item \code{volume} (string)
\item \code{pages} (string)
\item \code{doi} (string)
}
}
\examples{
data(cvdata)
pubs <- cvdata$publications$articles
insert_pubs(pubs)
}
\seealso{
\code{\link[=insert_pubs_web]{insert_pubs_web()}}, \code{\link[=insert]{insert()}}, \code{\link[=insert_talks]{insert_talks()}}, \code{\link[=insert_list]{insert_list()}}
}
