% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.R
\name{insert_talks}
\alias{insert_talks}
\title{Insert talks into your CV}
\usage{
insert_talks(data)
}
\arguments{
\item{data}{A list in the format expected by the \code{qmdcv} package. See
\code{data(example_data)} for an example.}
}
\value{
Markdown text to be rendered with Quarto.
}
\description{
\code{insert_talks()} converts a structured list into Quarto-style
Markdown text to be rendered as a list of talks.
}
\details{
Note: the code chunk in which this function is run should be given
the option \code{output: asis}
}
\examples{
data(example_data)
talks <- example_data$talks
insert_talks(talks)
}
\seealso{
\code{\link[=insert]{insert()}}, \code{\link[=insert_pubs]{insert_pubs()}}, \code{\link[=insert_list]{insert_list()}}
}
