% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcv_insert.R
\name{qcv_insert}
\alias{qcv_insert}
\title{Insert a CV entry}
\usage{
qcv_insert(data)
}
\arguments{
\item{data}{A list in the format expected by the \code{qcvutils} package. See
\code{data(qcvexample)} for an example.}
}
\value{
Literal markdown text and HTML code to be rendered with Quarto
}
\description{
\code{qcv_insert} converts a structured list (typically imported
from a YAML file) into literal markdown text and HTML code to be rendered
with Quarto. Note that the code chunk in which this function is run should
be given the \code{output: asis} option.
}
\examples{
data(qcvexample)
edu <- qcvexample$education
qcv_insert(edu)

}
