% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcv_insert.R
\name{qcv_insert}
\alias{qcv_insert}
\title{Insert entries into your CV}
\usage{
qcv_insert(data)
}
\arguments{
\item{data}{A list in the format expected by the \code{qmdcv} package. See
\code{data(qcvexample)} for an example.}
}
\value{
Markdown text to be rendered with Quarto.
}
\description{
\code{qcv_insert()} converts a structured list into Quarto-style
markdown text to be rendered as a CV section.
}
\details{
Note: the code chunk in which this function is run should be given
the option \code{output: asis}
}
\examples{
data(qcvexample)
edu <- qcvexample$education
qcv_insert(edu)
}
\seealso{
\code{\link[=qcv_insert_pubs]{qcv_insert_pubs()}}, \code{\link[=qcv_insert_talks]{qcv_insert_talks()}}, \code{\link[=qcv_insert_list]{qcv_insert_list()}}
}
