% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcv_insert.R
\name{qcv_insert_list}
\alias{qcv_insert_list}
\title{Insert a list into your CV}
\usage{
qcv_insert_list(data, type = "u")
}
\arguments{
\item{data}{A list in the format expected by the \code{qcvutils} package. See
\code{data(qcvexample)} for an example.}

\item{type}{Type of list to insert. Choose from among:
\itemize{
\item \code{"u"}: unordered list
\item \code{"1"}: numbered list
\item \code{"a"}: alphabetical list
\item \code{"n"}: no bullets
}}
}
\value{
Markdown text to be rendered with Quarto.
}
\description{
\code{qcv_insert_list()} converts a structured list into Quarto-style
markdown text to be rendered as a list.
}
\details{
This function is currently known to provide undesirable output when
supplied \code{title} and/or \code{details} are long AND \code{start}/\code{end}/\code{years} are
provided. When using this function, one should ideally supply only short
\code{title} and \code{details}. If longer \code{title} and/or \code{details} are needed,
consider using \code{\link[=qcv_insert]{qcv_insert()}}.

Note: the code chunk in which this function is run should be given
the option \code{output: asis}
}
\examples{
data(qcvexample)
teaching <- qcvexample$teaching
qcv_insert_list(teaching)
}
\seealso{
\code{\link[=qcv_insert]{qcv_insert()}}, \code{\link[=qcv_insert_pubs]{qcv_insert_pubs()}}, \code{\link[=qcv_insert_talks]{qcv_insert_talks()}}
}
